requirements.txt
gspread
gspread-dataframe
python-dotenv
beautifulsoup4